backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pandas==3.0.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
pydantic==2.12.5
python-multipart==0.0.22
alembic==1.18.4
python-dotenv==1.0.0